separate linear algebra class
Move QPSolver into this class and separate some matrix functions
