+++ cmd/lialg/qpsolver_test.go
@@ -1,0 +1,1 @@
+package lialg_test
