+++ cmd/lialg/rbfkernel_test.go
@@ -1,16 +1,16 @@
-package datamanager_test
+package lialg_test
 
 import (
 	"math"
 	"testing"
-	"transductive-experimental-design/cmd/datamanager"
+	"transductive-experimental-design/cmd/lialg"
 )
 
 func TestRBFKernelBasic(t *testing.T) {
 	x := []float64{1, 1}
 	y := []float64{2, 2}
 
-	value := datamanager.RbfKernel(x, y, 1)
+	value := lialg.RbfKernel(x, y, 1)
 	expected := math.Exp(-1)
 
 	// check if floating point error is within bounds
