--- cmd/qpsolver/qpsolver_test.go
@@ -1,1 +1,0 @@
-package qpsolver_test
